Renamed xlsxBuilder to commonServicesToolkit (makes more sense)
